refactor meeting state items into seperate files
