+++ packages/communication-react/review/communication-react.api.md
@@ -1253,9 +1253,8 @@ export interface MeetingAdapter extends AdapterState<MeetingAdapterState>, Adapt
 
 // @alpha
 export interface MeetingAdapterClientState extends Pick<CallAdapterClientState, 'devices'> {
-    displayName: string;
-    latestErrors: MeetingErrors;
-    meeting: MeetingState;
+    displayName: string | undefined;
+    meeting: MeetingState | undefined;
     userId: CommunicationIdentifier;
 }
 
@@ -1345,29 +1344,17 @@ export type MeetingCompositeProps = {
 // @alpha
 export type MeetingEndReason = CallEndReason;
 
-// @alpha (undocumented)
-export type MeetingErrors = unknown;
-
 // @alpha
 export type MeetingEvent = 'participantsJoined' | 'participantsLeft' | 'meetingEnded' | 'isMutedChanged' | 'callIdChanged' | 'isLocalScreenSharingActiveChanged' | 'displayNameChanged' | 'isSpeakingChanged' | 'messageReceived' | 'messageSent' | 'messageRead' | 'error';
 
 // @alpha
-export interface MeetingParticipant extends Pick<ChatParticipant, 'shareHistoryTime'>, Pick<RemoteParticipantState, 'displayName' | 'state' | 'videoStreams' | 'isMuted' | 'isSpeaking'> {
+export interface MeetingParticipant extends Pick<RemoteParticipantState, 'displayName' | 'state' | 'videoStreams' | 'isMuted' | 'isSpeaking'> {
     id: CommunicationIdentifier;
-    meetingEndReason: MeetingEndReason;
+    meetingEndReason?: MeetingEndReason;
 }
 
-// @alpha
-export interface MeetingState extends Pick<CallState, 'callerInfo' | 'state' | 'isMuted' | 'isScreenSharingOn' | 'localVideoStreams' | 'transcription' | 'recording' | 'transfer' | 'screenShareRemoteParticipant' | 'startTime' | 'endTime'>, Pick<ChatThreadClientState, 'chatMessages' | 'threadId' | 'properties' | 'readReceipts' | 'typingIndicators' | 'latestReadTime'> {
-    id: string;
-    meetingEndReason: MeetingEndReason;
-    participants: {
-        [key: string]: MeetingParticipant;
-    };
-    participantsEnded: {
-        [keys: string]: MeetingParticipant;
-    };
-}
+// @public
+export type MeetingState = unknown;
 
 // @public (undocumented)
 export type Message<T extends MessageTypes> = {
