+++ packages/react-composites/src/composites/MeetingComposite/state/MeetingCompositePage.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+/**
+ * Page state the Meeting composite could be in.
+ * @alpha
+ */
+export type MeetingCompositePage = 'configuration' | 'meeting' | 'error' | 'errorJoiningTeamsMeeting' | 'removed';
