+++ packages/react-composites/src/composites/MeetingComposite/state/MeetingEndReason.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { CallEndReason } from '@azure/communication-calling';
+
+/**
+ * Describes the reason the meeting ended.
+ * @alpha
+ */
+export type MeetingEndReason = CallEndReason;
